+++ Arrangement_on_surface_2/examples/Arrangement_on_surface_2/circular_line_arcs.cpp
@@ -72,7 +72,8 @@ int main() {
   Point_location _pl(arr);
   for (ArcContainer::const_iterator it = ac.begin(); it != ac.end(); ++it) {
     //insert(arr,_pl,*it);
-    insert(arr, *it, _pl);
+    //DONOTCOMMIT
+    //~ insert(arr, *it, _pl);
   };
 
   return 0;